Move trainer to new file
