+++ DeFRCNTrainer.py
@@ -1,0 +1,43 @@
+import os
+
+from detectron2.data import MetadataCatalog
+from detectron2.utils.events import CommonMetricPrinter, JSONWriter
+
+from defrcn.engine import DefaultTrainer
+from defrcn.evaluation import DatasetEvaluators
+
+
+class Trainer(DefaultTrainer):
+
+    @classmethod
+    def build_evaluator(cls, cfg, dataset_name, output_folder=None):
+        if output_folder is None:
+            output_folder = os.path.join(cfg.OUTPUT_DIR, "inference")
+        evaluator_list = []
+        evaluator_type = MetadataCatalog.get(dataset_name).evaluator_type
+        if evaluator_type == "coco":
+            from defrcn.evaluation import COCOEvaluator
+            evaluator_list.append(COCOEvaluator(dataset_name, True, output_folder))
+        if evaluator_type == "pascal_voc":
+            from defrcn.evaluation import PascalVOCDetectionEvaluator
+            return PascalVOCDetectionEvaluator(dataset_name)
+        if len(evaluator_list) == 0:
+            raise NotImplementedError(
+                "no Evaluator for the dataset {} with the type {}".format(
+                    dataset_name, evaluator_type
+                )
+            )
+        if len(evaluator_list) == 1:
+            return evaluator_list[0]
+        return DatasetEvaluators(evaluator_list)
+
+    def build_writers(self):
+        """
+        Minor change to get rid of default TensorboardXWriter
+        """
+        # Assume the default print/log frequency.
+        return [
+            # It may not always print what you want to see, since it prints "common" metrics only.
+            CommonMetricPrinter(self.max_iter),
+            JSONWriter(os.path.join(self.cfg.OUTPUT_DIR, "metrics.json")),
+        ]
